[Splitting of files][patient.py > form_generator.py -- utils.py][event.js (provider code) -> provider_operations.js] and related fixes
