+++ phr/templates/includes/share_phr.js
@@ -26,7 +26,7 @@ $.extend(SharePhr.prototype,{
 		this.bind_controller()
 		//console.log(["ss",me.args['event_id']])
 		this.render_folder_section(args['event_id'],args['method'])
-		Events.prototype.get_linked_providers(this.args['profile_id'])
+		ProviderOperations.prototype.get_linked_providers(this.args['profile_id'])
 	},
 	bind_controller: function(){
 		var me = this;
