+++ phr/templates/pages/provider_page.py
@@ -1,7 +1,7 @@
 import frappe
-from phr.templates.pages.patient import get_data_to_render
+from phr.templates.pages.form_generator import get_data_to_render
 from phr.phr.phr_api import get_response
-from phr.templates.pages.patient import get_base_url
+from phr.templates.pages.utils import get_base_url
 import json
 import datetime
 from frappe.utils import getdate, date_diff, nowdate, get_site_path, get_hook_method, get_files_path, \
